rename: 네이밍 컨벤션 변경, 라우팅과 연관 없는 컴포넌트 폴더 이동 - components 폴더 생성: 라우팅과 연관없는 컴포넌트들 보관 - 네이밍 컨벤션 변경 - 폴더 내에 index가 있으면 컴포넌트 명과 동일, PascalCase 적용 - 폴더 내에 index가 없으면 snake_case 적용
