+++ app/[id]/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import BreweryDetailsProps from "@/types/BreweryDetailsProps";
 import BreweriesApi from "@/services/BreweryService";
-import BreweryForm from "./brewery_edit_form";
+import BreweryEditForm from "@/components/BreweryEditForm";
 
 interface Props {
   params: { id: string };
@@ -11,5 +11,5 @@ export default async function BreweryDetails({ params }: Props) {
   const breweriesApi = new BreweriesApi();
   const id = params.id;
   const brewery: BreweryDetailsProps = await breweriesApi.fetchBreweryById(id);
-  return <BreweryForm brewery={brewery} />;
+  return <BreweryEditForm brewery={brewery} />;
 }
