--- app/new/BreweryDetails.module.scss
@@ -1,14 +1,0 @@
-.form {
-  width: 100vw;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  row-gap: 20px;
-}
-
-.info_box {
-  display: flex;
-  flex-direction: column;
-  width: 50vw;
-}
