+++ src/components/BreweryList/BreweryCard/index.tsx
