+++ src/components/SearchBar/SearchBar.module.scss
