+++ src/components/icons/BeerIcon.tsx
