+++ src/components/icons/MoonIcon.tsx
