+++ tsconfig.json
@@ -17,13 +17,12 @@
     "baseUrl": ".",
     "paths": {
       "@/app/*": ["app/*"],
-      "@/pages/*": ["pages/*"],
-      "@/styles/*": ["styles/*"],
-      "@/mocks/*": ["__mocks__/*"],
+      "@/components/*": ["src/components/*"],
       "@/services/*": ["src/services/*"],
       "@/types/*": ["src/types/*"],
       "@/utils/*": ["src/utils/*"],
-      "@/hooks/*": ["src/hooks/*"]
+      "@/hooks/*": ["src/hooks/*"],
+      "@/mocks/*": ["__mocks__/*"]
     },
     "plugins": [
       {
